requirements.txt
mysql-connector-python==8.0.33
pymongo==4.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
